move contant definitions for mongoose models from resolvers back to mongoose models to dont repeat code
